pbmt: optimize code & bump coupledL2
